fix: code formatting
